pkg/automerge: do not add ready-to-merge label if changes requested
If a user has requested for changes and made a GH comment afterwards for
that PR the automerge logic would assume it would be OK to mark the PR
ready-to-merge. Following the same logic applied by GH, a PR will only
be marked as ready-to-merge if the user explicitly approved the
requested changes after requesting for changes.

Signed-off-by: André Martins <andre@cilium.io>