client: unassign PR reviewer on approval
Users marked for requested reviews get assignments for the PR. Once they
submit a review, and this review is an approval of the pull request,
they are supposed to unassign themselves from the PR. This helps
Janitors finding PRs that need their attention. But removing the
assignment is currently done manually, and users often omit it.

Let's automate unassignment on review, when the review concludes with an
approval of the PR.

Signed-off-by: Quentin Monnet <quentin@isovalent.com>