pkg/jenkins: Adapt wording for potential flakes
When a test fails and MLH can't find a related flake issue, it states
that the flake hasn't been seen before. Unfortunately, MLH isn't always
able to find the flake issues even if they exist. This commit therefore
adapts the wording to not give the impression that no flake issue
exists.

Reported-by: Bruno Miguel Custódio <brunomcustodio@gmail.com>
Signed-off-by: Paul Chaignon <paul@cilium.io>